Вынести общие базовые классы выражений в parsers/anum_core.py
Устраняет массивное дублирование базовых классов AST (Symbol, Connection,
AssociativeRoot, ConnectionForm и др.) между anum_prover.py и
mtc_formula_prover.py. Оба доказателя теперь импортируют их из общего
модуля anum_core. Поведение сохранено; mtc_formula_prover.py сократился
до 1232 строк (ниже лимита CI 1500).
